[MCP slack-bot] Migration
